chore: migrate retrieval to ADR-099 slice convention
Pilot 1 of the Vertical-Slice-Architektur convention (ADR-099):
move the `retrieval` slice to the canonical layout
`src/drift/retrieval/{contracts,handlers,mcp,tests}.py` and verify
that no behaviour, public import path, or MCP tool schema changes.

Changes in this commit:
- src/drift/retrieval/contracts.py: re-export public models (initial
  minimal contract surface, ADR-099 Pflicht-Datei).
- src/drift/retrieval/handlers.py: thin delegations onto RetrievalEngine
  (retrieve, cite, get_engine, invalidate_cache).
- src/drift/retrieval/mcp.py: full copy of the legacy
  `mcp_router_retrieval.py` (run_retrieve / run_cite).
- src/drift/mcp_router_retrieval.py: replaced by a thin re-export shim
  `from drift.retrieval.mcp import *` so existing importers keep
  working (FR-007, SC-002).
- src/drift/mcp_server.py: switch the two lazy imports to the new
  slice path (drift.retrieval.mcp).
- src/drift/retrieval/__init__.py: broaden `__all__` with
  RetrievalEngine, build_corpus, BM25Index, tokenize, clear_engine_cache,
  cite, get_engine, invalidate_cache, retrieve.
- src/drift/retrieval/tests/__init__.py: enable pytest discovery.
- tests/test_retrieval_corpus.py -> src/drift/retrieval/tests/test_corpus.py
- tests/test_retrieval_search.py -> src/drift/retrieval/tests/test_search.py
- tests/test_mcp_retrieval_tools.py -> src/drift/retrieval/tests/test_mcp.py
  (only change in tests: `repo_root` fixture now uses `parents[4]`
   instead of `parent.parent` because tests live deeper).
- specs/003-vsa-pilot1-retrieval/: spec, plan, research, data-model,
  contracts, quickstart, checklists, tasks.
- docs/decisions/ADR-099-vertical-slice-architecture-convention.md.

Verification:
- `pytest` (default discovery): 6719 passed, 132 skipped.
- `pytest src/drift/retrieval/tests/`: 28 passed.
- `mypy src/drift/retrieval/ --strict --exclude tests/`: 0 errors.
- `python -m drift mcp --schema` pre/post: identical (SC-008).
- `grep` for `from drift.{intent,calibrate,signals,scoring,ingestion,output}`
  in `src/drift/retrieval/`: 0 hits (SC-006).
- `from drift.retrieval.search import RetrievalEngine`: ok (SC-007).

Decision: ADR-099
