+++ specs/003-vsa-pilot1-retrieval/checklists/requirements.md
@@ -1,0 +1,41 @@
+# Specification Quality Checklist: VSA-Pilot 1 — Migration `retrieval` → Slice-Konvention
+
+**Purpose**: Validate specification completeness and quality before proceeding to planning
+**Created**: 2026-04-28
+**Feature**: [spec.md](../spec.md)
+
+## Content Quality
+
+- [x] No implementation details (languages, frameworks, APIs)
+- [x] Focused on user value and business needs
+- [x] Written for non-technical stakeholders (soweit bei einem Entwickler-internen Refactor möglich)
+- [x] All mandatory sections completed
+
+## Requirement Completeness
+
+- [x] No [NEEDS CLARIFICATION] markers remain — eine offene Option (A vs. B für MCP-Router) ist bewusst an `/clarify` delegiert, nicht als NEEDS-CLARIFICATION-Marker
+- [x] Requirements are testable and unambiguous
+- [x] Success criteria are measurable
+- [x] Success criteria are technology-agnostic (keine Frameworks, Datenbanken o. ä. genannt)
+- [x] All acceptance scenarios are defined
+- [x] Edge cases are identified (Circular Import, fehlende Chore-Vorbedingung, models.py-Kollision)
+- [x] Scope is clearly bounded (Out of Scope explizit und nicht verhandelbar)
+- [x] Dependencies and assumptions identified (Chore-PR, Boundary-Audit, Test-Strategie)
+
+## Feature Readiness
+
+- [x] All functional requirements have clear acceptance criteria (FR-001 bis FR-010)
+- [x] User scenarios cover primary flows (Layout, MCP-Kanal, Tests, Import-Garantie)
+- [x] Feature meets measurable outcomes defined in Success Criteria (SC-001 bis SC-008)
+- [x] No implementation details leak into specification
+
+## Open Items
+
+- [ ] **Option A vs. B für MCP-Router**: Muss via `/clarify` entschieden werden, bevor Planung startet.
+- [ ] **`models.py` vs. `contracts.py`**: Wird im PR-Review entschieden; kein Blocker für Spec-Freigabe.
+- [ ] **Chore-Vorbedingung** (`pyproject.toml` testpaths, codecov.yml): Muss als separater PR vor oder parallel zum Pilot-PR abgeliefert werden.
+
+## Notes
+
+- Alle items außer den oben genannten offenen Punkten sind erfüllt.
+- Die Spec ist bereit für `/clarify` (Option A/B) und danach für `/plan`.
