+++ specs/003-vsa-pilot1-retrieval/data-model.md
@@ -1,0 +1,209 @@
+# Data Model: VSA-Pilot 1 — `retrieval` Slice
+
+**Generated by**: `/speckit.plan`
+**Date**: 2026-04-28
+
+---
+
+## Slice Layout nach Migration
+
+```
+src/drift/retrieval/
+├── __init__.py          # Public API — breiter __all__ (Q5-Entscheidung)
+├── cache.py             # Unverändert — persistenter Corpus-Cache
+├── contracts.py         # NEU — Slice-Vertragstypen (Initial: leer/minimal)
+├── corpus_builder.py    # Bleibt als internes Hilfsmodul (parse_* Funktionen)
+├── fact_ids.py          # Unverändert — stabile Fact-ID-Generierung
+├── handlers.py          # NEU — konsolidierte Use-Case-Logik (RetrievalEngine-Wrapper)
+├── index.py             # Unverändert — BM25-Index-Implementierung
+├── mcp.py               # NEU (verschoben aus mcp_router_retrieval.py)
+├── models.py            # Unverändert — frozen Pydantic Domain-Models
+├── search.py            # Bleibt als internes Hilfsmodul (RetrievalEngine + Cache-Logik)
+└── tests/
+    ├── __init__.py      # Leer
+    ├── test_corpus.py   # Verschoben aus tests/test_retrieval_corpus.py
+    ├── test_mcp.py      # Verschoben aus tests/test_mcp_retrieval_tools.py
+    └── test_search.py   # Verschoben aus tests/test_retrieval_search.py
+
+src/drift/mcp_router_retrieval.py  # Re-Export-Shim (bleibt temporär)
+```
+
+---
+
+## Bestehende Domain-Models (`models.py` — unverändert)
+
+### `FactChunk`
+
+```python
+class FactChunk(BaseModel):
+    model_config = ConfigDict(extra="forbid", frozen=True)
+    fact_id: str        # Stable structured identifier (ADR-091)
+    kind: str           # "policy" | "roadmap" | "adr" | "audit" | "signal" | "evidence"
+    source_path: str    # Posix-relative path within repo
+    line_start: int     # 1-based inclusive
+    line_end: int       # 1-based inclusive
+    text: str           # Verbatim chunk content
+    sha256: str         # SHA-256 hex digest of chunk text
+    tags: tuple[str, ...]  # Structured tags (signal_id, adr_number, section)
+```
+
+**Relationships**: Used by `BM25Index`, `RetrievalEngine`, `CorpusManifest`
+
+### `RetrievalResult`
+
+```python
+class RetrievalResult(BaseModel):
+    model_config = ConfigDict(extra="forbid", frozen=True)
+    fact_id: str
+    kind: str
+    source_path: str
+    line_range: tuple[int, int]
+    excerpt: str        # Truncated chunk text
+    sha256: str
+    tags: tuple[str, ...]
+    score: float        # BM25 relevance score
+```
+
+**Relationships**: Output of `RetrievalEngine.retrieve()`
+
+### `CorpusManifest`
+
+```python
+class CorpusManifest(BaseModel):
+    model_config = ConfigDict(extra="forbid", frozen=True)
+    schema_version: int         # CORPUS_SCHEMA_VERSION constant
+    generated_at: str           # ISO 8601 timestamp
+    repo_root: str              # Absolute path
+    corpus_sha256: str          # Hash of all chunks deterministically
+    sources: tuple[SourceEntry, ...]
+```
+
+### `SourceEntry`
+
+```python
+class SourceEntry(BaseModel):
+    model_config = ConfigDict(extra="forbid", frozen=True)
+    path: str       # Posix-relative source path
+    sha256: str     # File content hash
+    chunks: int     # Chunk count from this source
+```
+
+---
+
+## Neue Datei: `contracts.py` (minimal, initial leer)
+
+Zweck: Pflicht-Erfüllung ADR-099 + Erweiterungspunkt für künftige Slice-Vertragstypen.
+
+```python
+"""Retrieval slice contracts — ADR-099.
+
+Slice-lokale DTOs und Vertragstypen. Das bestehende `models.py` enthält
+die Domain-Models (FactChunk, RetrievalResult etc.) und bleibt unverändert.
+
+Neue Slice-Vertragstypen (z. B. für externe API-Bindings) werden hier
+definiert, sobald sie entstehen.
+"""
+
+# Initial leer — Pflicht-Datei nach ADR-099 §3
+```
+
+---
+
+## Neue Datei: `handlers.py` (Use-Case-Konsolidierung, Q4-Entscheidung B)
+
+Zweck: Zentrale Use-Case-Logik des Slices. Konsolidiert `RetrievalEngine`-Lifecycle
+und `build_corpus` als öffentliche Slice-Einstiegspunkte.
+
+### Haupt-Entitäten in `handlers.py`
+
+| Symbol | Quelle | Funktion |
+|---|---|---|
+| `retrieve(path, query, top_k, kind, signal_id)` | Use-Case delegiert an `RetrievalEngine.retrieve` | Kern-Retrieval-Use-Case |
+| `cite(path, fact_id)` | Use-Case delegiert an `RetrievalEngine.cite` | Fact-Zitation |
+| `get_engine(repo_root)` | Wrapper um `RetrievalEngine.for_repo` | Engine-Lifecycle |
+| `invalidate_cache(repo_root)` | Wrapper um `clear_engine_cache` | Cache-Invalidierung |
+
+**Wichtig**: `RetrievalEngine` selbst verbleibt in `search.py` als Implementierungsdetail;
+`handlers.py` ist der Use-Case-Layer (thin delegation auf `search.py`). Die Q4-Entscheidung
+„Konsolidierung" bedeutet hier: klare Use-Case-Funktionen als Einstiegspunkt, nicht
+physische Code-Bewegung von `RetrievalEngine`-Internals.
+
+---
+
+## Neues Modul: `mcp.py` (verschoben aus `mcp_router_retrieval.py`)
+
+### MCP-Tool-Funktionen (identisch zum Pre-Migration-Stand)
+
+| Funktion | MCP-Tool | Parameter |
+|---|---|---|
+| `run_retrieve(path, query, top_k, kind, signal_id)` | `drift_retrieve` | `path: str`, `query: str`, `top_k: int`, `kind: str \| None`, `signal_id: str \| None` |
+| `run_cite(path, fact_id)` | `drift_cite` | `path: str`, `fact_id: str` |
+
+**Validierungsregeln** (aus `mcp_router_retrieval.py` übernommen, unverändert):
+- `query` darf nicht leer sein → `DRIFT-RAG-001`
+- `top_k` muss ≥ 1 sein → `DRIFT-RAG-002`
+- `kind` muss in `{"policy", "roadmap", "adr", "audit", "signal", "evidence"}` liegen → `DRIFT-RAG-003`
+- `path` muss existieren → `DRIFT-RAG-004`
+- `fact_id` muss gefunden werden → `DRIFT-RAG-005`
+
+---
+
+## `__init__.py` Public Surface nach Migration (Q5-Entscheidung B)
+
+```python
+__all__ = [
+    # fact_ids
+    "MigrationRegistry",
+    "generate_adr_id",
+    "generate_audit_id",
+    "generate_evidence_id",
+    "generate_policy_id",
+    "generate_signal_id",
+    # models (Domain)
+    "CorpusManifest",
+    "FactChunk",
+    "RetrievalResult",
+    "SourceEntry",
+    # search / engine
+    "RetrievalEngine",
+    "clear_engine_cache",
+    # corpus_builder
+    "build_corpus",
+    # index
+    "BM25Index",
+    "tokenize",
+]
+```
+
+---
+
+## Re-Export-Shim: `mcp_router_retrieval.py`
+
+```python
+# Re-export shim — temporär bis alle Konsumenten auf drift.retrieval.mcp migriert sind.
+# Entfernen in einem Folge-PR nach vollständiger Konsumenten-Migration.
+from drift.retrieval.mcp import *  # noqa: F401,F403
+```
+
+---
+
+## State Transitions / Lifecycle
+
+```
+Corpus nicht vorhanden
+    → load_or_build(repo_root)          [cache.py]
+    → build_corpus(repo_root, sources)  [corpus_builder.py]
+    → BM25Index(chunks)                 [index.py]
+    → RetrievalEngine._is_fresh check   [search.py]
+    → Cache in _engines[repo_root]
+
+Query
+    → RetrievalEngine.retrieve(query, top_k, kind, signal_id)
+    → BM25Index.search(query_tokens, top_k)
+    → list[RetrievalResult]
+
+Cite
+    → RetrievalEngine.cite(fact_id)
+    → FactChunk lookup via _by_id dict
+    → FactChunk | None
+```
