+++ specs/003-vsa-pilot1-retrieval/research.md
@@ -1,0 +1,169 @@
+# Research: VSA-Pilot 1 — Migration `retrieval` → Slice-Konvention (ADR-099)
+
+**Generated by**: `/speckit.plan`
+**Date**: 2026-04-28
+**Spec**: [spec.md](spec.md)
+
+All NEEDS CLARIFICATION items from the Technical Context are resolved here.
+All five open decisions were resolved in Clarifications (Session 2026-04-28) before planning.
+
+---
+
+## Decision 1: MCP-Router-Kanal (Q1)
+
+**Decision**: Option A — `mcp_router_retrieval.py` → `src/drift/retrieval/mcp.py` + Re-Export-Shim
+
+**Rationale**: Vollständige Slice-Self-Containment gemäß ADR-099; MCP-Kanal lebt im Slice.
+Der Shim am alten Pfad (`from drift.retrieval.mcp import *  # noqa: F401,F403`) hält
+Abwärtskompatibilität bis alle Konsumenten migriert sind. `mcp_router_retrieval.py`
+ist der **einzige** externe Konsument des Slices — verifiziert per Boundary-Audit
+(`work_artifacts/vsa_pilot1_retrieval_boundary_audit.md`).
+
+**Alternatives considered**:
+- Option B (In-Place-Delegation): Verworfen, weil Slice unvollständig wäre;
+  widerspricht ADR-099-Geist der vollständigen Self-Containment.
+
+**Mypy-Hinweis**: `__all__` in `mcp.py` definieren, damit Mypy beim Wildcard-Re-Export
+keine Warnung erzeugt (`implicit reexport`-Modus nicht aktiv, aber defensive Praxis).
+
+---
+
+## Decision 2: `models.py` vs. `contracts.py` (Q2)
+
+**Decision**: Option C — Koexistenz: `contracts.py` (neu, Slice-Verträge/DTOs) +
+`models.py` (behalten, Domain-Models)
+
+**Rationale**: Minimale Änderung, minimales Risiko. `models.py` enthält bewährte, frozen
+Pydantic-Modelle (`FactChunk`, `RetrievalResult`, `CorpusManifest`, `SourceEntry`), die
+von anderen Retrieval-Modulen intern verwendet werden. Eine Umbenennung würde alle internen
+Imports (`from drift.retrieval.models import ...`) brechen ohne Mehrwert im Pilot-PR.
+
+`contracts.py` wird neu als leere (oder minimal kommentierte) Datei angelegt — Pflicht-Erfüllung
+der ADR-099-Konvention; kann in einem späteren PR mit Slice-Vertragstypen befüllt werden.
+
+**Alternatives considered**:
+- Option A (Rename): Zu viele Importe zu ändern; kein Pilot-PR-Scope.
+- Option B (Neu anlegen mit echten Verträgen): Overengineering für Pilot; handlers.py
+  wurde zur Use-Case-Konsolidierung gewählt (Q4), nicht contracts.py.
+
+---
+
+## Decision 3: Chore-PR-Sequenz (Q3)
+
+**Decision**: Option A — Chore-PR (testpaths + codecov) muss vor Pilot-PR gemergt sein
+
+**Rationale**: Pytest-Discovery von `src/drift/retrieval/tests/` setzt `testpaths = ["tests", "src/drift"]`
+in `pyproject.toml` voraus. Ohne diesen Eintrag findet `make check` keine Slice-Tests;
+der Pilot-PR wäre nicht eigenständig grün verifizierbar.
+
+Chore-PR-Inhalt:
+```toml
+# pyproject.toml [tool.pytest.ini_options]
+testpaths = ["tests", "src/drift"]
+```
+```yaml
+# codecov.yml — Coverage-Pfade
+coverage:
+  include:
+    - src/drift/*/tests/**
+```
+
+Validierung: Dummy-Slice-Test anlegen → `make check` grün → Dummy entfernen → Chore-PR mergen.
+
+**Alternatives considered**:
+- Option B (Kombinieren): Einzel-PR würde zwei orthogonale Änderungen mischen;
+  erschwert Bisect bei CI-Fehlern.
+- Option C (TODO-Kommentar): Pilot-PR wäre ohne Test-Co-lokalisierung unvollständig
+  per ADR-099-Definition.
+
+---
+
+## Decision 4: `handlers.py`-Granularität (Q4)
+
+**Decision**: Option B — Konsolidierung: Use-Case-Logik aus `search.py` und `corpus_builder.py`
+wird nach `handlers.py` gezogen; alte Module werden zu internen Hilfsdateien
+
+**Rationale**: `handlers.py` als Konsolidierungspunkt entspricht der ADR-099-Absicht
+(Use-Case-Logik im Slice gebündelt). `search.py` enthält `RetrievalEngine` und
+`clear_engine_cache` — diese Use-Case-Logik wandert in `handlers.py`.
+`corpus_builder.py` enthält `build_corpus` als Haupt-Entry-Point — dieser wandert ebenfalls.
+
+**Konkrete Zuordnung**:
+
+| Modul (heute) | Wandert nach | Verbleibt als |
+|---|---|---|
+| `search.py::RetrievalEngine` | `handlers.py` (oder delegiert von dort) | internes Hilfsobjekt |
+| `search.py::clear_engine_cache` | `handlers.py` | — |
+| `corpus_builder.py::build_corpus` | `handlers.py` | internes Helper |
+| `corpus_builder.py::parse_*` | verbleiben in `corpus_builder.py` | interne Hilfsfunktionen |
+
+**Wichtig**: `RetrievalEngine` muss weiter über `drift.retrieval.search.RetrievalEngine`
+importierbar bleiben (SC-007). Backward-compat via `__init__.py` re-export (Q5).
+
+**Alternatives considered**:
+- Option A (Thin delegation): Hätte `handlers.py` als leeren Wrapper gelassen;
+  kein echter Konsolidierungsnutzen.
+- Option C (Skeleton only): Kein Mehrwert; Pilot-PR hätte `handlers.py` ohne Inhalt.
+
+---
+
+## Decision 5: `__init__.py` Public Surface (Q5)
+
+**Decision**: Option B — Breite Public Surface: alle aktuell genutzten public Symbole
+aus `search.py`, `corpus_builder.py`, `index.py` in `__all__`
+
+**Rationale**: Der heutige `__init__.py` exportiert bereits einen breiten Symbol-Satz
+(bestätigt per `read_file`):
+
+```python
+# Aktueller __init__.py — bereits vorhanden:
+from drift.retrieval.fact_ids import (
+    MigrationRegistry, generate_adr_id, generate_audit_id,
+    generate_evidence_id, generate_policy_id, generate_signal_id,
+)
+from drift.retrieval.models import (
+    CorpusManifest, FactChunk, RetrievalResult, SourceEntry,
+)
+from drift.retrieval.search import RetrievalEngine
+```
+
+Nach Konsolidierung (Q4) muss `__init__.py` zusätzlich `clear_engine_cache` und
+`build_corpus` exportieren (falls diese öffentlich bleiben), damit externe Konsumenten
+nicht brechen.
+
+**`__all__` nach Migration** (vollständig):
+```python
+__all__ = [
+    # fact_ids
+    "MigrationRegistry", "generate_adr_id", "generate_audit_id",
+    "generate_evidence_id", "generate_policy_id", "generate_signal_id",
+    # models
+    "CorpusManifest", "FactChunk", "RetrievalResult", "SourceEntry",
+    # search / handlers
+    "RetrievalEngine", "clear_engine_cache",
+    # corpus_builder
+    "build_corpus",
+    # index
+    "BM25Index", "tokenize",
+]
+```
+
+**Alternatives considered**:
+- Option A (Nur RetrievalEngine): Zu eng; bricht potenzielle Konsumenten von `build_corpus`,
+  `BM25Index` etc.
+- Option C (Kein __all__): Schlechtere Mypy- und IDE-Unterstützung; widerspricht
+  ADR-099-Konvention expliziter Slice-API.
+
+---
+
+## Architecture Confirmation
+
+Aus dem Boundary-Audit und Code-Read verifiziert:
+
+| Eigenschaft | Status |
+|---|---|
+| Kein ausgehender Import auf andere `drift.*`-Slices | ✅ bestätigt |
+| Einziger externer Konsument: `mcp_router_retrieval.py` Line 11 | ✅ bestätigt |
+| Alle Dateien self-contained (stdlib + `drift.retrieval.*` intern) | ✅ bestätigt |
+| POLICY §18 Audit-Pflicht ausgelöst | ❌ NEIN — kein Signal/Scoring/Output-Eingriff |
+| Commit-Typ `chore:` korrekt | ✅ — reine Strukturmigration |
