+++ src/drift/retrieval/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Drift Retrieval slice tests (ADR-099)."""
