+++ src/drift/retrieval/tests/test_mcp.py
@@ -14,7 +14,8 @@ from drift.retrieval.search import clear_engine_cache
 
 @pytest.fixture(scope="module")
 def repo_root() -> Path:
-    return Path(__file__).resolve().parent.parent
+    # ADR-099: tests live in src/drift/retrieval/tests/, repo root is 4 levels up.
+    return Path(__file__).resolve().parents[4]
 
 
 @pytest.fixture(autouse=True)
